ADD: Buy Lotteries Path and config database

+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/repository/UserTicketRepository.java
@@ -3,6 +3,6 @@ package com.kbtg.bootcamp.posttest.repository;
 import com.kbtg.bootcamp.posttest.model.UserTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserTicketRepository extends JpaRepository<UserTicket, Long> {
+public interface UserTicketRepository extends JpaRepository<UserTicket, Integer> {
 
 }
